producer/requirements.txt
kafka-python==2.1.0
requests==2.31.0

